add baseclass for easier inheritance
