+++ spec/victor/svg_base_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe SVGBase do
+  it "does not define #method_missing" do
+    expect{ subject.polygon }.to raise_error NoMethodError
+  end
+end
